Add base test utils.
