+++ packages/core/package.json
@@ -46,7 +46,7 @@
     "exit": "^0.1.2",
     "fast-glob": "^2.2.6",
     "fs-extra": "^7.0.1",
-    "i18next": "^13.1.1",
+    "i18next": "^13.1.2",
     "json5": "^2.1.0",
     "lodash": "^4.17.11",
     "os-locale": "^3.1.0",
